Modified and added scripts
Modified and added scripts for the CPU paddle and to keep track of score.
CPU paddle now has basic AI to follow the ball's movement at a somewhat reasonable speed.
Scores are now tracked and displayed. For each score the ball resets to the center.
The ball trajectory seems to keep the same trajectory from before the score which is a bug.
The top and bottom borders only bounce the ball sometimes which is another bug to fix later.
Going to add menu and game over screens next.